pass user-created 'tags' to underlying buck rule in python_binary() infra macro
Summary:
Infra macros for python_binary() buck rule type currently do not pass `tags` to the underlying buck rule unless the rule is a test.
This diff passes macro tags down as `labels` to the buck rule regardless if this is a test. This is useful for querying buck to find python binaries with specific user-created tags.

Reviewed By: philipjameson

Differential Revision: D8538699

fbshipit-source-id: 51b32e932fc052ee8b29d41acded1b48eb9a521f