Teach 'extend_build_mode' about 'compiler'
Summary:
The 'extend_build_mode' macro was added so as to allow child
build modes to use nearly all of the options specified by a
parent build mode, but add options to one or two of them.

Teach this 'extend_build_mode' macro about the new 'compiler'
option that was recently added to 'create_build_mode'. It doesn't
make sense to "extend" a compiler option, which is a string. So
instead, have the child override the parent when it specifies its
own compiler option.

Reviewed By: philipjameson

Differential Revision: D8732457

fbshipit-source-id: b579c77f4e1b2ff6783a12fab1a5e19429002098