Fix issue with root-owned .pyc files
Summary:
Buck running as non-root users lacks permission to clean up `__pycache__` directories, which can cause subsequent builds to fail.

We could set PYTHONBYTECODEPATH / PYTHONPYCACHEPREFIX, but this opens a can of worms about which scratch space to use, how to clean it up, etc.

So, for the time being, let's take a perf hit and avoid the problem altogether.

Differential Revision: D8897890

fbshipit-source-id: 55a713ceb3c33e54eb421e72a5c2f8f736564982