Rearrange demo_sendstreams.py to get ready for `libbtrfsutil`
Summary:
The main gist of this change is as follows:

- Before: `demo_sendstreams.py` is a library that's used from `python_binaries`, as well as a binary that is run directly through system `python3`. This is confusing, and requires `libbtrfsutil` to be available both on the system, and through Buck.
- After:
-- `demo_sendstreams.py` is the thin library used nearly as before.
-- `make_demo_sendstreams.py` is a `python_binary`, no longer executed by system Python.

Also remove the executable bit from a bunch of the `.py` files to make it more clear that they are intended to be run via `python_binary`, rather than directly.

We still likely have a few things that run under system python via `-m` (such as the `print_gold_demo_sendstreams` script that I drive-by renamed in this diff). But we can hunt them down later on a one-off basis.

Reviewed By: zeroxoneb

Differential Revision: D8930485

fbshipit-source-id: bce35f21613358ec9a285b2e2dee138ded72edd5