Cleaner image construction -- sudo at the leaves
Summary:
The tl;dr here is that I get much cleaner control flow from running nearly all the business logic as the repo user, and only using `root` for the leaf filesystem actions, all through one class.

Specifically, `Subvol` eliminates a bunch of confusing workarounds (see the next diff for their actual deletion), and also lets me drastically simplify the main image-building logic in `compiler` (a few diffs later in the stack).

You have probably noticed that I have been churning on this test code for the last several diffs. That's because it is a good, fast, simple test-bed for what I want to do next with the compiler itself. Specifically, the `btrfs_util` module here (not to be confused with osandov's `libtrfsutil`) will become a standalone library with proper tests, and I will then use it inside `compiler`.

Moving the abstraction boundary to the leaves does come with a couple of downsides — however, I feel that they are justified by the simpler code:
 - This adds a few dozen `sudo` calls to the `demo_sendstream` test, which adds about 450ms to its runtime. We will see the same sort of per-sudo penalty of about 10ms when building production images. If an image requires 1000s of small steps, that adds up to a lot of wasted time. Of course, optimizations are possible, including a persistent privileged server, etc, but this does have worse perf out of the box.
 - It's a step back on integrating `libbtrfsutil`. I like `btrfsutil`, but using it requires the mass of the logic to run as root -- crossing privilege boundaries without leaving Python-land is, of course, not possible. If we ever go with a "RPC to a server running as root" for optimization reasons, we can also switch most of the btrfs functionality to `btrfsutil` at that time. For now, I want the simpler code flow.

If I may dream about the future, we will probably eventually want to replace the `sudo` shell-outs with one of:
 - a privileged server that calls out to `btrfsutil`, persists for the duration of the build, and only allows access to the subvolumes being built,
 - `guestfs`
Doing the work at the leaves affords us the luxury of changing the implementation later, without incurring major surgery.

Reviewed By: zeroxoneb

Differential Revision: D9005868

fbshipit-source-id: b4dc8f017b42b992e959bd77d3d793ebc859b653