TempSubvolumes owns its temporary directory
Summary: This makes it easier to write tests that make subvolumes (takes common logic out of `demo_sendstreams.py`). It also lets `TempSubvolumes` ensure that the subvolumes have sane permissions on the path to the subvolumes its owns (see the comment at the end of `get_volume_for_current_repo`).

Reviewed By: marcinpe

Differential Revision: D9112070

fbshipit-source-id: 2d79bd92abf7ad52a0779a9bbdfe8471a8155c76