buck macros: add module compilation rule for `module.modulemap`s
Summary:
This adds a new rule which compiles a group of headers with a `module.modulemap`
file into a `.pcm` file.  This is meant as an alternative to clang's automatic
local module caching framework, and allows Buck to remotely cache `.pcm` files.

The interface allows headers to be specified as either a directory or a map of
headers, and so can be plugged into existing `cpp_library` and `cpp_library_external`
rules to automatically build modules in a Buck-cache friendly way.

Note: no effort has yet been put into making sure compiled modules are deterministic
when using this mechanism, and it does appear to embed absolute paths into the resulting
PCM files.

Reviewed By: philipjameson

Differential Revision: D8955772

fbshipit-source-id: 9860f5b768d30728ce8dd501c178d2d41d035f28