compiler: build command for tarball item
Summary:
This is the first of a few changes that will introduce `build` implementations of the actions formerly serialized via `build_command`. At the end, I can remove `build_command`, as well as `get_image_build_command_target`.

The goal of this series of diffs is to cut the dependency on the separate image-building tool. There are a few benefits here:
 - Build times become several times faster because we lose almost all dependencies
 - Self-contained code with fewer abstraction hops is easier to understand
 - This will be much closer to working as an OSS utility

Reviewed By: marcinpe

Differential Revision: D9164674

fbshipit-source-id: d3812dd94f8603b6ae803f4241eef1bf93f1f2d9