compiler: build command for copy-file item
Summary:
See D9164674 for the overarching motivation.

Specifically for this change, besides what it says on the tin, I added explicit support for both symbolic and octal `mode` values. In the past, we would accept symbolic strings, but the end result would not be well-defined, since it would depend on the initial permissions of the file or directory (possibly involving umask, etc). So, I changed that.

Reviewed By: marcinpe

Differential Revision: D9174752

fbshipit-source-id: d9a6ea3c812653599bf83cc9b6ff68ddafc8d3e7