enable overriding platform for wheels
Summary: The idea behind is: you want to build a par for Windows (or another platform possibly) but you're building on Linux. For buck, the platform is Linux and so it's going to use the Linux versions of wheels. Overriding the platform for wheels lets us keep everything going the Linux way, but use the Windows wheels for third-party dependencies.

Reviewed By: oxo42

Differential Revision: D9267632

fbshipit-source-id: c1e8270c249b9288117fdc792b1bcb8b4d31b3db