thrift grammar rules: discard original tainted files
Summary: Previously `sed -i.bak` was used to rename the yacc-/bison- generated file `foo` to `foo.bak`, then filtered and written to `foo` again.  The `.bak`s are no longer "used" but since they are present in the `gen` dir, they're still used when hashing the output of the `genrule`, leading to lots of unnecessary building.

Reviewed By: mzlee

Differential Revision: D9499874

fbshipit-source-id: 467df1e0fef624bdfb1052c3cb3166aa0aa4f815