Add src and dep helpers
Summary:
This adds a simple lib for some of our naming helpers. Right now this just moves source_name from base.py. It will eventually have some of the things like 'get_all_deps' and the like in it. Better names are welcome.

This wasn't used in base.py since it has isinstance methods and the like gating a lot of its usage. This logic hasn't changed for a while, so it should be fine in the intermediary while we're migrating rules.

Reviewed By: andrewjcg

Differential Revision: D9304697

fbshipit-source-id: fc9b1135e8a2329d502c6e976bf40b54c8b8e936