Update the naming scheme for garbage-collectable subvols
Summary:
This diff commits both `subvolume_garbage_collector` and `SubvolumeOnDisk` to using the same subvolume naming scheme, and adds tests.

The old scheme was `<name>:<version>/<name>`, and it was only enforced by the GC, which means that `SubvolumeOnDisk` could create non-GC-able subvolumes.

The new scheme is `<rule name>:<version>/<subvol>`. It is enforced both by GC and by `SubvolumeOnDisk`.

This is necessary because we must allow rules to build subvolumes with names that differ from the rule name. A specific use-case: the subvolume name may contain its own versioning, which should not be reflected in the rule name.

The consequence of more relaxed naming is that (for the time being) the subvolume must be the ONLY filesystem object inside of the temporary wrapper directory.

I also simplified the interface of `SubvolumeOnDisk`, and rearranged its flow slightly to eliminate one unnecessary validation pass.

Reviewed By: marcinpe

Differential Revision: D9550073

fbshipit-source-id: 97dff2f408297be40f4e7d70a3bf9ff087e7406f