Clowny ducks quack int, not float
Summary:
Q: Why?
A: Because a Python update decided that at least one of these needs to be an int.

Q2: Why is not in a base class?
A2: Because the pain of creating a shared dependency for all the unit tests doesn't seem worthwhile, given that this copy-pasta is very limited and easily grepped.

Q3: Why 12345?
A3: Greppability

Reviewed By: marcinpe

Differential Revision: D9550184

fbshipit-source-id: 10a0b9fbf1c76b4d0f80906a860c46f5159121f7