reduce link time for gcc LTO builds
Summary:
Allow the number of partitions to increase beyond
the default of 32, which is more appropriate for our huge
binaries and massively parallel systems.

Before this change, the hhvm lto link phase looked like:

  ld x 1 = 3m..
    lto1-wpa x 1 = 6m20s
    lto-wrapper x 1 = 20s
    ld x 1 = 1m
    lto1-ltrans x 32 = 2m10s .. 23m
  ld x 1 = ..13m

Notice the bad distribution of 2m .. 23m running time
of the lto1-ltrans processes, due to a couple of
very large partitions being processed.

This sets the number of partitions to 2x the base lto setting,
thus giving a worst case lto1-ltrans of 13m.

I.E. this improves the hhvm LTO link time by about 10m.

Other parameters (new in platform007) to consider in this area in future are:
`lto-min-partition` and `lto-max-partition`.

Reviewed By: markw65

Differential Revision: D9629142

fbshipit-source-id: f98d2396be1b331c6dd5a9a0119e2487773b4f7d