Fix dependency on macros from infra_macros/macro_lib/convert/container_image
Summary:
Everything under `container_image` was getting picked up by the `**/*.py` glob, even though only two of the `*.py` files are intended to be Buck macros.

This was causing the build to fail, since many of the files in the container_image directory are python 3 only and do not compile with python 2. These errors are ignored in dev mode builds, but in dbg and opt builds the par generation fails if there are syntax errors in any of the files.

The fix is to exclude `container_image` from the glob, and to add its 2 macro files manually, via a dependency.

Reviewed By: simpkins

Differential Revision: D9728617

fbshipit-source-id: c58c271bc65852eb71f7e91d636b6a679dee33f4