Ensure mac test runs on macro changes
Summary:
Recently when a change to the macros broke Java builds on laptops,
this test started failing, but was not triggered by Sandcastle so
it didn't prevent the change from landing. This diff sets it up in
a way that any change to the macro library will trigger this test.

Reviewed By: sazonovkirill

Differential Revision: D9793896

fbshipit-source-id: b003e1a9e7986bbb6e334539727f78db9d6571c8