Silently ignore BUILD_MODE global_compiler overrides
Summary:
Three design decisions interact in a way that makes it difficult to override the
"opt" compiler:

1. The 'global_compiler' setting cannot be overridden, if it is attempted,
   Buck asserts.
2. Build mode overrides applied to "opt" are applied to sanitizer modes like
   "opt-asan" as well.
3. The opt-asan build mode sets its global_compiler to Clang.

In D9411889, an attmept was made to add build mode overrides for "opt" that
set the compiler to GCC for services that are not migrating to Clang yet.
As a result of that change and (1) from the list above, any opt-asan build
for the services overridden would fail due to a Buck assert.

Loosen the assertion behavior in order to make the compiler override easier
to use. With this diff, if a BUILD_MODE file attmepts to override a globally
set compiler, the file override is silently ignored.

Reviewed By: igorsugak

Differential Revision: D9931181

fbshipit-source-id: 9ea2e9eab5571a5c85094a037a36a4d577526aeb