Fix the HHH crash
Summary:
The problem is to do with retention of CAFs (top-level expressions),
which need careful handling when doing dynamic linking. We were
already doing this in sigma_server, but not in HHH.

See comments for details. Because this requires starting the GHC
runtime with some special config options, we have to define our own
`main` function in C.

Reviewed By: watashi

Differential Revision: D9934536

fbshipit-source-id: 174c748f6c7f92f1e3f27bd3f28fa6b6caeb2507