Adds import mem analyzer to buck python.helpers.
Summary:
The import_time_mem.py script already existed, but takes a few steps to get working. My goal here was to make this script easier to use.

I've broken import_time_mem into two parts- generate_par_imports and analyze_par_imports, and left the original import_time_mem in place so I don't break anyone's TARGETS files that currently imports it.

I added new rules to the python buck macros file, so we can generate new par files that can be used to easily run the import memory analysis. There is a gen-rule, that runs the generate_par_imports file and outputs an imports_file.py that can be used for analysis.

The analyze-imports.par can run with no arguments and it will automatically use the generated imports_file. If it is passed an argument, it will use that filename instead. The generate-imports.par file can be run alone to just generate the imports_file.

To run the whole sequence, a user needs to:

1. Add `analyze_imports = True` to their TARGETS file binary.

2. Run `buck build <project> -c python.analyze_imports=True`

3. Run `buck-out/.../<project_name>-analyze-imports.par`

And you will get an output of the biggest RSS and VSZ offenders in the imports.

Reviewed By: ambv

Differential Revision: D9694378

fbshipit-source-id: cdca0c2ec3a0cb272aa5cbe6e392e5618ff2705d