Commit semantics for storage abstraction
Summary:
tl;dr -- this makes repo snapshots effectively "all or nothing" in terms of the persistent data they emit. This is not true transactionality, but it is fairly robust auto-cleanup.

We're writing blobs to `Storage`, but the storage IDs of those blobs are going into a DB, where commits can and will be aborted.

This change adds a mandatory & explicit `commit` method to `StorageOutput`, with the additional feature that committed blobs can be auto-removed if the `Storage.writer()` context has an uncaught exception (via `output.commit(remove_on_exception=True)`).

The new semantics makes it easy for my snapshot code not leave behind orphan blobs when something goes wrong with the snapshot.

Reviewed By: marcinpe

Differential Revision: D9733810

fbshipit-source-id: d37e8153d86b9098b4d303ade4b7f882f5266829