Move plugin logic from Storage to reusable base
Summary:
I need another set of plugins for DBs to allow switching between SQLite for dev/OSS and MySQL for prod.

Having a single base class provide the plugin functionality ensures that both types of plugins have a consistent UI.

This also adds an `add_argparse_arg` for adding auto-documented command-line options that instantiate plugins. This is exercised by the CLIs that I add in later diffs, but I include an example of the `--help` output in the test plan.

Reviewed By: marcinpe

Differential Revision: D10123227

fbshipit-source-id: f6add130d8ca09faf0135d09acaaf3883cc392a6