Remove remaining global duplicate finder excludes
Summary:
All targets where those excluded classes still conflict either have
duplicate finder disabled (primarily third-party-java) or they have
an explicit exclude.

`module-info.class` is left for now because it does not have any
effect when building with Java <9 and when we start moving to
building with Java >=9, it is not clear whether we will want to use
the module info or remove it and ignore it.

Differential Revision: D10116530

fbshipit-source-id: 0a1374d3298b41f9a89f3fc83a21273f918c9fa1