Context for managing a DB connection (SQLite/MySQL)
Summary:
For testing & local development, SQLite is super-convenient -- it's faster to read & write than remote MySQL due to lack of connection handshakes, it's easy to reset (`rm`), I can have multiple copies, unit-testing is a breeze.

For production data, we of course want a proper, managed & replicated MySQL DB.

So in this diff, and a later one for the DAOs, I add a bit of boilerplate to paper over the differences between the two DB engines.  In terms of connection handling

Reviewed By: marcinpe

Differential Revision: D10123386

fbshipit-source-id: a204bb60216ce28a2374af7a7054e7d20bf06429