convert/go: Add automatic target for benchmarks from go_unitest rules
Summary:
Since benchmarks are built into the Go's tests we can generate a
simple command_alias automatically to invoke tests defined in _test.go files
without an additional rule.

This could be optional, not a lot of Benchmarks exist in go_unitest targets
today, but it really doesn't hurt adding the rule for every `go_unittest` and
standardization here is better than configurability.

Reviewed By: mkaczanowski

Differential Revision: D10184481

fbshipit-source-id: c575e7e920fed17292473c842782417ebb02c4c4