Estimate total disk usage of repos
Summary:
This is the only inessential part of the system, in that we don't HAVE to know how big the snapshots are. But it's interesting, and a very useful sanity check.

Sizing one repo is easy enough, you just add up the sizes of the objects. In a multi-repo setting, the same objects can -- and do -- live in different repos. Ideally, we wouldn't double-count those.

So, we should de-duplicate by checksums. To make matters more interesting, different repos can, and do, use different checksum algorithms.

For RPMs, we always compute and record a canonical checksum (sha384), which enables us to dedupe them effectively. The motivation for canonical checksums is so that we can detect when the same RPM name has different content in different repos (again, something that does occur).

The final wrinkle is that for performance, snapshots can be sharded by the modulo of a hash of an RPM name. In that case, for RPMs outside of our shard, we will NOT have the canonical SHA384. But we might still have e.g a SHA256. And we might see a copy in the current shard of the same object with both a SHA256 (from repodata) and SHA384 (computed canonical). For this reason, the `RepoSizer` code merges synonym sets. It's pretty easy, and sets my conscience at ease.

Repodata objects don't have canonical checksums. If there are duplicates of the same repodata with different checksums, we'll just double-count those, but it's likely negligible -- at FB, repodata is about 0.04% of the repo size.

Reviewed By: marcinpe

Differential Revision: D10190084

fbshipit-source-id: 8f3eac0571d0e95c8bcacd0d1454e40dfae3336a