Parse primary repodata from SQLite or XML
Summary:
RPM repo updates ought to be "atomic" in the sense that we upload new repodata, new RPMs, and then swap the `repomd.xml` that points at the aforementioned objects.

To take advantage of this atomicity, we discover the RPMs to snapshot by parsing the "primary" repodata object that was mentioned by `repomd.xml`.

`createrepo` by default emits both `-primary.sqlite.bz2` and `-primary.xml.gz`, but in the wild, there exist other variations:
 - no SQLite (**ahem** EPEL, why would you do that?)
 - SQLite compressed with .gz (yay, faster decompression, now how about `.zst` for fast **and** compact?)

As detailed in the docblocks, parsing XML is really slow, so we prefer SQLite just to keep development time sane.
But all of the above cases do come up, so we have to handle them.

Reviewed By: marcinpe

Differential Revision: D10190126

fbshipit-source-id: ab3ab62ec196fa3eee6ba8e2bd7432f2bdcb5690