buck macros: add `compile_with_modules` parameter to opt-out from modules
Summary:
This adds a parameter for C++ rules to allow them to opt-out of using modules
when compiling their `.cpp` sources, as a way to workaround one-off issues
(e.g. compiler ICE).  Note that rules will still compile their headers into a
module (for dependent rules that use modules to consume), even with this flag
set to `False`.

Reviewed By: vitaut

Differential Revision: D9322000

fbshipit-source-id: cc9765f0ef3568e6f284f5bcb4667eb207971726