Repo server to serve a JSON snapshot from VCS over HTTP
Summary:
This is the integration point for image builds. Specifically,
 - the repo snapshot code (in later diffs) produces `RepoSnapshot`s, which are committed to version control
 - the image compiler:
    - creates an `AF_INET` / `SOCK_STREAM` socket
    - starts `repo_server.py` with that socket
    - arranges via `LISTEN_FDS` for `systemd-nspawn` to pass that socket to our `--private-network` container
    - arranges for the nspawn container to bind that socket to a known port
    - bind-mounts a fake `yum.conf` pointing at the socket's port on localhost (inside the container's network namespace)
    - runs `yum install` inside the nspawn container

This server doesn't do much. It deserializes a bunch of repo's `RepoSnapshot`s, and serves them via HTTP. For blobs in `Storage`, it reads the blobs, and performs checksum validation on the client's behalf. The point here is that I don't trust the storage to necessarily be immutable (errors, bitrot, if not outright malice), so validating-on-fetch is a sound insurance policy. The source-controlled checksums ought to be a bit harder to mess up than mostly cold data.

Reviewed By: marcinpe

Differential Revision: D10190348

fbshipit-source-id: f3702842bd7c534480fa00fb980b86f0bd4c60cc