Add to_label and target_to_label methods
Summary:
Adds methods to convert RuleTarget structs into buck labels.

Note that this does not preserve previous logic that allowed for 'unknown cells' that was part of the OSS effort. That will be handled in a different way in the future. The repo can either be: 'None', third-party/third-party-tools, or an arbitrary string.

Reviewed By: andrewjcg

Differential Revision: D9951500

fbshipit-source-id: 500f5b83bd75f59503f0980ce58e8e867247a27e