Per-target C++ compiler override
Summary:
EDIT: This actually seems like the right approach. Original diff summary follows.

I think this is probably the wrong approach, but I thought I'd send it up as a
strawman for now. I think its behavior might be close to what I want, which is
to set the mode/opt compiler to GCC for specific targets. As an example in
this diff, I removed the override for the fbcode/zippydb directory, and instead
exclude just the biggest CPU consumer in that directory, //zippydb/server:zippydb_server.

I want to do this because my queries from N44718 show that 125 C++ targets in
fbcode that are built with mode/opt account for most of gCPU. I'd like for
just these 125 targets to "override" the default mode/opt compiler, Clang, and
use GCC instead. Then, for the rest of the half, I'll slowly migrate these
remaining 125 one by one.

Let me know if this approach is somewhere along the right track. If not, here's
another idea: is there a way to override the `fbcode.default_compiler=`
Buck config option for these specific targets somehow? Perhaps by adding local
.buckconfig files or something?

Reviewed By: igorsugak

Differential Revision: D10081164

fbshipit-source-id: b2bfef632c9c65717c63a440376b0c9c979022d1