Update `yum.conf` rewriting logic to (sort of) "isolate" yum from the host
Summary:
The previous `modify_repo_configs` was a stub for the logic that I actually want. The goal of manipulating a snapshot's `yum.conf` is to preserve the overall intent of the configuration, but at the same time to:
 - point all the URLs at our `repo-server` working against a snapshot
 - eliminate various rough edges of `yum` where it tries to interact with the host filesystem -- it should only be able to read the snapshot, and touch its `installroot`.

In practice, complete isolation of `yum` cannot be achieved purely through configuration changes. The `yum-from-snapshot` diff describes the problems, and includes a sandbox for `yum` based on Linux namespaces + bind-mounts.

Reviewed By: marcinpe

Differential Revision: D10374645

fbshipit-source-id: a06fb835ce6765fd1110eb8900c6083edd364723