Fix build_repos.py not to make varying RPMs with the same name
Summary:
I built a snapshot of a set of test repos and got "mutable RPM" errors because, indeed, I was making RPMs with the same name, but different build times.

This change replaces duplicate RPMs with symlinks to the original.

The updated repos will be committed in the next diff.

Reviewed By: marcinpe

Differential Revision: D10374625

fbshipit-source-id: 70863cc09e4c250e9ae46f0de7bf3766d5aa918c