Temporarily committing a test repo snapshot
Summary:
For testing `yum-from-snapshot`, I need a test snapshot. I generated this from the auto-generated `rpm/tests/repos` (which is reasonable to commit since it involves `createrepo` and other external dependencies) using the as-yet-uncommitted `snapshot-repos`.

This really shouldn't be committed to VCS long-term because it's more meaningful to exercise `rpm/tests/repos` -> `snapshot` -> consumers with an up-to-date version of `snapshot-repos`. The snapshot process is pretty fast, and this eliminates a bunch of manual work for maintaining `snapshot.tgz` as the code evolves. This integrated approach will also give extra exercise to various snapshot-consumer integrations.

Reviewed By: marcinpe

Differential Revision: D10374643

fbshipit-source-id: 0403ff7498076506b3bdc20bf3e58234158e5783