Stop overriding crypto wheels to unbreak platform007/tupperware
Summary:
Python wheel overrides were enabled for platform007 as part of D10131417,
while the cryptography and asn1crypto wheels were added to the overrides
list in D10362351.  Turns out, the cryptography wheel is not correctly building
for py3-platform007, thereby breaking the Tupperware build when platform007
was enabled for overrides.

This removes asn1crypto and cryptography from the overrides list, allowing
use of the TP2 projects for platform007, and unbreaking Tupperware.

unbreakmeow

(Note: this ignores all push blocking failures!)

Reviewed By: mkatsevVR

Differential Revision: D10391369

fbshipit-source-id: 88e0bfbfbb920e5f7034b6eb3afe68c0a2118b4b