Add a `yum-from-test-snapshot` utility to support testing Buck image builds
Summary:
When we build an image with RPMs, we need to tell the image compiler how to install its RPMs.

This is the interface with the image compiler. It is basically `yum`, but with a deliberate three-extra-dashes breakage in compatibility, with the goal of forcing people who might try to use `yum` directly to think at least a moment.

The other reason for interface incompatibility is that I have another capability in mind I will likely need to bolt on to this CLI, and it being NOT `yum` makes that easy.

Reviewed By: zeroxoneb

Differential Revision: D10416711

fbshipit-source-id: 2a57b3f44dda0e26a34e964f2a429aa7ac6df534