Add a check for logging library deps
Summary:
This diff extends duplicate finder and makes it check whether
logging library implementation classes are present in the
classpath. This will help us enforce that logging library
implementations are only dependencies of Java binaries where they
are less likely to conflict with each other.

For now no actual implementation classes are forbidden since the
rule would fail for many libraries in the repository. Actual
logging library implementation classes will be added to the rule
in future diffs after the repository is cleaned up a bit.

This diff also extends the scope of duplicate_finder beyond the
original intent and beyond its name. The plan is to rename
duplicate_finder to classpath_auditor in a future diff.

Differential Revision: D10414825

fbshipit-source-id: 978f86b83804ddac629fcf5016914a9676301e2b