Handle large classpaths
Summary:
Duplicate finder would fail to run for targets with many entries on
the classpath because the classpath argument would exceed the OS's
argument length limit. This diff fixes the issue by using arg4j's
"at-file" syntax which lets us pass arguments through a dynamically
generated file by prefixing it with `@` on the command line. Note
that the OS's argument length limits do not apply to `echo` because
it is built into the shell and not an external process.

Differential Revision: D10439916

fbshipit-source-id: 47e8260402c000811f41f9d44f03f238252441c1