Add unpacked `tests/snapshot/` to replace `snapshot.tgz`
Summary:
In D10374643, I added this temporary snapshot as a packed `.tgz`. Later, I realized that `.tgz` is not a good model for accessing production snapshots (perf cost, complexity), so I fixed the main blocker in fbcode `python_library` for doing it "right": D10455994.

This new snapshot still has a workaround for a problem with our buck macros, in that `gpg_keys/` contains a `placeholder`, since we cannot build PARs/XARs with empty directories. This isn't a problem in practice, since prod snapshots always have GPG keys, and the test one will, too, before long.

Reviewed By: zeroxoneb

Differential Revision: D10470396

fbshipit-source-id: 971167faba885903e10c7e6e4426584817d0b53b