Detect RPM action conflicts (e.g. install/remove)
Summary: Since I cannot make `yum` do the work of detecting install/remove conflicts by failing on non-existent removes, I want to block this explicitly. To be clear: in the absence of this check, `PhaseOrder` would still resolve conflicts in a deterministic fashion, but then people might begin to rely on the behavior, and we could never ever change phase orders again. Having this behavior be well-defined would mostly just cause grief for users, because it's not normally what you mean.

Reviewed By: marcinpe

Differential Revision: D10477363

fbshipit-source-id: 331a562322ae82fce44fe97e670a1ba8c65850d5