Allow `ImageItem`s to depend on image changes made by phases (e.g. RPMs)
Summary:
Note the change in `parent_layer` in `compiler/tests/TARGETS`. I am, somewhat reasonably, trying to copy a tarball into a directory that was created by an RPM.

Prior to this change, this kind of operation would fail, since the `ParentImageLayer` or `FilesystemRootLayer` used in evaluating the `ImageItem` dependency graph would not know anything about the state of the post-RPM filesystem.

With this change, we build the phases before validating the `ImageItem` dependency graph, which allows them to depend on the results of phase builds.

Reviewed By: zeroxoneb

Differential Revision: D10483314

fbshipit-source-id: 427766e3978a358beefae5d8064907b5a3c05f4c