Version logging is inversed when building for windows.
Summary:
Currently:

`We're showing this warning because you're building for py3-win_amd64 and the default version of pyyaml for this platform (4.1) doesn't match the default version for py3-gcc-5-glibc-2.23 (None). The resulting binary might not work on py3-win_amd64. Make sure there is a py3-win_amd64 wheel for None version of pyyaml.`

In reality, we don't have any windows version matching the linux one.

Reviewed By: philipjameson

Differential Revision: D10469704

fbshipit-source-id: 5fbea8f04f4157be1cae466cdba2c1989062ab43