fbcode tools: asm_filter source + core-tool listing
Summary:
A first pass at adding a new component to the toolchain.  This is tricky because the tool itself needs to be compiled.  (In contrast to other Python scripts, which do consume a bit too much CPU time as it is.)

I wrote this is as one big cpp file, with no includes and no dependencies to libraries, except for standard C++ (11) itself, glibc, etc., very basic things.

Let me know if I did the bzl parts right :-p

Reviewed By: andrewjcg

Differential Revision: D10466877

fbshipit-source-id: 21961b6f086e78263ebc136ce139a1cb98b52a77