Fix a few instances of exe/location confusion in Buck macros
Summary:
This was not previously documented, but for linktree PARs `$(location ...)` uses the linktree symlink as the artifact -- and that never changes. This means that updates to the referenced Python sources would never trigger a rebuild, resulting in source consistency issues.

In contrast `$(exe ...)` does know to rebuild if the source files backing a linktree PAR had changed.

Reviewed By: marcinpe

Differential Revision: D10855613

fbshipit-source-id: 49555ba00595aaf6dc08044c65a69a1a01640225