Use lists as a default for labels param
Summary:
These two functions have default values of None and `()` (empty tuple) for label defaults.

I'm surprised Buck actually accepts an explicit `None` as a value for labels when it expects a list. Other macros generally dont though, as they expect labels to either be a list or missing

For the tuple, AFAICT no one in the codebase uses a tuple to define their labels. Labels are also a param that are often mutated (specifically - they get new additions), so throwing an immutable value there messes with things.

This diff replaces those defaults with a default of (effectively) `[]`.

Reviewed By: philipjameson

Differential Revision: D12559268

fbshipit-source-id: 5473199c4b273624123f0a12410873e7ccff4678