buck macros: auto-detect and recover from module non-determinism
Summary:
We're currently seeing some rare case of non-determinism causing broken modules
to get generated.  As it's proving hard to repro locally, this diff had the module
rules auto-detect and recover from this non-determinism (by building the module
twice)  This is meant as a temporary workaround to avoid user interruption while investigating
the issue.

Reviewed By: pixelb

Differential Revision: D12817013

fbshipit-source-id: 6ac3b0fc3094933badf8fc339b80574eb1d00198