Fix flavored Python targets not picking up versioned resources.
Summary:
The parameters

    versioned_lang_ppflags
    versioned_shared_lib
    versioned_header_dirs
    versioned_static_lib
    versioned_static_pic_lib

of `cpp_library_external` Buck target type, and the parameters

    versioned_srcs
    versioned_resources

of `python_xxx` Buck target types allow certain resources to be associated
with specific versions of Python. Most of these parameters are used in
third-party-buck projects with auto-generated TARGETS file (see
`buckify_tp2.py`) that only include a declaration for the base non-flavored
Python versions, and therefore flavored Python targets would not be able to
pickup these versioned resources in their dependencies. This is an oversight
from D9696599, and we fix it in this diff.

Reviewed By: aaronkroth

Differential Revision: D10443365

fbshipit-source-id: 50015ba3d9109dac931a73b24d8cf7bb3c38dac8