Allow hack to omit whls on some platforms
Summary:
Allow a python_wheel to explicitly be set to None to signal that it is not available on a platform.
We do not generate whl s for urls set to None, and we do not add them as platform deps. You can still include the main python_library target, but for unsupported platforms, you aren't going to get anything. If a wheel_platform is set, then the _override_wheel code will make sure that for all available buck platforms (that were specified as None), that we point to the overridden whl.
For the default wheel, we require you to specify a version, even if you're not using it. Again, the non-matching platform_deps will make sure that you're getting either the right whl, or no wheel.

Reapplication of D12890768

Reviewed By: cooperlees

Differential Revision: D12897257

fbshipit-source-id: c443c0505318fe318877aca2413a2f6f5fdaaa9b