Refactor global compiler check in order to remove sanitizers dependency on build mode overrides
Summary:
The fact that sanitizers (transitively) depend on build modes overrides posed a problem when we attempted to enable fbandroid/native sanitizers - in particular, windows (!!!) does not support symlinks, so any build mode override symlinks would break windows xplat compilation.

While the work to refactor build mode implementation continues, it appears that the sanitizers dependency on build modes is minimal, and really incidental.  So, refactor the two methods that were used by sanitizers into standalone macros so that we can use those, without pulling in the rest of the dependencies.

Reviewed By: andrewjcg

Differential Revision: D12942942

fbshipit-source-id: 8add162444a9cbc152ed05d2d7f35411c3b68f45