buck macros: restrict C/C++ rule deps to only fbcode/third-party repos
Summary:
Currently, we rely on fbcode macros to implement a lot of our toolchain and so deps outside of
fbcode or third-party2, and onto different rule types in other cells and can cause issues (e.g. missing
sanitizer support, wrong/duplicate third-party libs).  Add an error to prevent this until we're ready to
open it up.

Reviewed By: igorsugak

Differential Revision: D12987301

fbshipit-source-id: 26c96d91d326e2ca0798bf30449c27c3f674615a